fix(ui): k_index float, double-click select-all, mouse drag selection
- SolarData.h / BandConditionsData.h: k_index int → float
- NOAAProvider: static_cast<int> → static_cast<float> for Kp parse
- SolarPanel: format K:%d → K:%.1f (narrow and wide layouts)
- SpaceWeatherPanel: format %d → %.1f; colorForK signature int → float
- WebServer: std::to_string(k_index) → snprintf %.1f for Kp text output
- SetupScreen onMouseUp: add double-click select-all for Identity and
  Spotting tab fields (matches existing Appearance tab pattern)
- TextInput: add onMouseMove (extends selection while mouse button held)
  and onMouseUp (clears drag state); mouseDown_ tracks drag in progress
- SetupScreen: override onMouseMove to forward drag to active TextInput
  using stored field rects; call ti->onMouseUp() on mouse release

+++ src/core/BandConditionsData.h
@@ -17,7 +17,7 @@ struct BandConditionsData {
   std::vector<BandStatus> statuses;
   std::chrono::system_clock::time_point lastUpdate;
   int sfi = 0;       // Solar Flux Index used for calculations
-  int k_index = 0;   // K-index used for calculations
+  float k_index = 0.0f; // K-index used for calculations
   bool valid = false;
 };
